+++ models/team.js
@@ -255,6 +255,14 @@ exports.createTeamForEvent = function(event, user, callback) {
   });
 };
 
+exports.getTeamForEvent = function (event, callback) {
+  const conditions = {
+    teamEventId: event._id
+  };
+
+  Team.findOne(conditions, callback);
+};
+
 // TODO: should this do something with ACL?
 exports.updateTeam = function(id, update, callback) {
   if (update.users) {
@@ -348,47 +356,4 @@ exports.removeUserFromAllAcls = function(user, callback) {
   update.$unset['acl.' + user._id.toString()] = true;
 
   Team.update({}, update, {multi: true, new: true}, callback);
-};
-
-/*
-Get a team based on event id.  This will return the team 
-directly associated with the event.
-*/
-exports.getTeamByEventId = function(eventId) {
-  var conditions = {};
-  if (eventId) {
-    conditions.teamEventId = {
-      '$eq': eventId
-    };
-  }
-
-  const promise = new Promise(function(resolve, reject) {
-    Team.findOne(conditions)
-      .exec()
-      .then(team => {
-        resolve(team ? team : undefined);
-      })
-      .catch(err => reject(err));
-  });
-
-  return promise;
-};
-
-exports.addUserWithPromise = function(team, user) {
-  var update = {
-    $addToSet: {
-      userIds: mongoose.Types.ObjectId(user.id)
-    }
-  };
-
-  const promise = new Promise(function(resolve, reject) {
-    Team.findByIdAndUpdate(team._id, update)
-      .exec()
-      .then(team => {
-        resolve(team ? team : undefined);
-      })
-      .catch(err => reject(err));
-  });
-
-  return promise;
 };